// mobile/src/data/graphql/API.ts
export type ModelStringInput = {
  ne?: string | null,
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  contains?: string | null,
  notContains?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  size?: ModelSizeInput | null,
};
// ...
export type ModelBooleanInput = {
  ne?: boolean | null,
  eq?: boolean | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
};
// ...
export type ModelIDInput = {
  ne?: string | null,
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  contains?: string | null,
  notContains?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  size?: ModelSizeInput | null,
};
// ...
export type ModelRACIMOFilterInput = {
  id?: ModelIDInput | null,
  Name?: ModelStringInput | null,
  LinkageCode?: ModelStringInput | null,
  Configuration?: ModelStringInput | null,
  createdAt?: ModelStringInput | null,
  updatedAt?: ModelStringInput | null,
  and?: Array< ModelRACIMOFilterInput | null > | null,
  or?: Array< ModelRACIMOFilterInput | null > | null,
  not?: ModelRACIMOFilterInput | null,
  _deleted?: ModelBooleanInput | null,
};
// ...
export type ModelStringKeyConditionInput = {
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
};

export enum ModelSortDirection {
  ASC = "ASC",
  DESC = "DESC",
}
// ...
export type ModelUVAFilterInput = {
  id?: ModelIDInput | null,
  latitude?: ModelStringInput | null,
  longitude?: ModelStringInput | null,
  altitude?: ModelStringInput | null,
  fields?: ModelStringInput | null,
  enabled?: ModelBooleanInput | null,
  createdAt?: ModelStringInput | null,
  userID?: ModelIDInput | null,
  racimoID?: ModelIDInput | null,
  updatedAt?: ModelStringInput | null,
  and?: Array< ModelUVAFilterInput | null > | null,
  or?: Array< ModelUVAFilterInput | null > | null,
  not?: ModelUVAFilterInput | null,
  _deleted?: ModelBooleanInput | null,
};
// ...
export type GetRACIMOQueryVariables = {
  id: string,
};
// ...
export type ListRACIMOSQueryVariables = {
  filter?: ModelRACIMOFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};
// ...
export type UVAsByRacimoIDQueryVariables = {
  racimoID: string,
  createdAt?: ModelStringKeyConditionInput | null,
  sortDirection?: ModelSortDirection | null,
  filter?: ModelUVAFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};